change odometry pipline

+++ src/main.cpp
@@ -9,8 +9,6 @@
 #include <opencv2/core/utility.hpp>
 
 #include "odometry.hpp"
-#include "odometry_func/odometry_frame.hpp"
-#include "odometry_func/odometry_settings.hpp"
 #include "tmp.hpp"
 
 using namespace cv;
@@ -26,7 +24,7 @@ int main(int argc, char** argv)
         0, fy, cy,
         0, 0, 1);
     //float depthFactor = 5000;
-    double depthFactor = 0.1;
+    float depthFactor = 0.1;
     Ptr<Scene> scene = Scene::create(true, frameSize, intr, depthFactor);
 
     std::vector<Affine3f> poses = scene->getPoses();
@@ -87,6 +85,7 @@ int main(int argc, char** argv)
 
     return 0;
 }
+
 
 /*
 {
